api/requirements.txt
Flask==2.2.2
kafka-python==2.0.2
pyspark==3.3.1
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
flask-cors==3.0.10
requests==2.28.1
pandas==1.5.1
numpy==1.23.3